Corrected Connection String, Added onto ViewModel
Corrected the connection string to connect to the database project. Modified UserDAL to use the correct Sql variables instead of MySql variables and changed the stored procedure call. Added onto the viewmodel implementation to correctly disable the log in button when the username and password fields are empty and added an error string if the credentials are incorrect. Made ViewModel implement INotifyPropertyChanged instead of ViewModelBase.