Fixed RegisterUser Tests, Modified OpenCreateAccount Test Implementation
Fixed the tests around the RegisterUser method to validate the error message returned for the exception thrown test case. Modified the test implementation of the OpenCreateAccount command to use the mock.VerifyGet() method. Although this test still does not pass, it provides a different implementation that can be modified to pass later on.