Added CreateLodging Form, Lodging Panel to TripOverview, and Starting ViewModels
Added the create lodging form and added command to trip overview for navigating to that form. Added starting CreateLodgingViewModel to be used with the form and LodgingViewModel to be used to populate the lodging panel in trip overview. Functionality will be added once Lodging backend is added. Renamed method to better fit functionality.