Made EndDate A Required Field for Events
Removed the nullable operator from EndDate in IEvent so that the end date is not null. As discussed in previous meetings, the end date for events will default to the end of the trip if one is not specified. Updated waypoint table in database to account for end date not being null.