Reworked ValidationManager::FindClashingEvent
Reworked the method to return an IList of the clashing events found and passed in another parameter for the event to exclude from finding clashing events, in the case of editing and not wanting to not consider the event that is being edited as a clashing event. Updated tests and references. Reworked the Equals method in Transportation and Waypoint to no longer account for fields that can change in editing. Updated Equals tests.